Further Prototype Fixes, Spider Boss, Siren Level Improvements
There are further general improvements.

A prototype version of the spider boss level has been created. An unfinished version of the spider boss has been created. Placeholder graphics are used.

The siren level is functional, although the level design should be improved. Placeholder graphics are used.

A proper game over function has been added. When this is called, the user is sent back to the level select screen.