Ship Level Transition Bug Fixed
The level transition bug that occurs when a player finishes the siren level has been fixed.

The game reads a map property named edgeMap to determine which map to warp to.

Map boundaries have been created for this level that do not cause a previous bug where the player cannot fly past the end of the screen.