Enemy + Combat Improvements
When the player hits the attack key, the game will continue the sword motion for a specified amonut of time (rather than the player killing all enemies upon contact when the attack key is held).

Enemies can be attacked multiple times. In previous versions of the game, the health would be lowered every frame (which would near-instantly kill an enemy over several frames if the enemy health was greater than the player's damage). To do this, the enemy is given a brief period of invulnerability (currently 0.5 seconds), which is currently being represented by the enemy being partially transparent.