King Aetios prototype
King Aetios has been added to the palace level.

He starts walking to the left when "_KINGAETIOSWALK":"" is added to the current dialogue entry.

The dialogueUpdate() function for each NPC is run when a dialogue box is rendered. This is currently used for the king to check if it is the appropriate time to walk over.