Spawn Portal After Boss Fight
If a portal has a 'spawnAfterBossBattle' property set to true, the portal will only appear if every boss is defeated. The spider boss level has been modified to demonstrate this.

spawnAfterSpiderFlower is used for the portal in the spider boss level to indicate that the portal should not spawn unless the player has picked up the flower.